Removendo antigo arquivo de texto onde se dava entrada dos repositório; Adicionando vericaçao ao tentar remover os repositorios de temp; Adicionado o example do .env

+++ src/repositories.js
@@ -104,5 +104,9 @@ export function createPullRequest(repoName, username) {
 
 export function clear() {
   console.log("Removendo diretórios temporários...");
-  shell.rm("-rf", "*");
+  const pathDirectoryList = (shell.pwd()).split("/");
+  const actualDirectory   = pathDirectoryList[pathDirectoryList.length - 1] ;
+  if (actualDirectory === "temp"){
+    shell.rm("-rf", "*");
+  }
 }
